Split config files
