+++ config/bootstrap_null.py
@@ -1,0 +1,2 @@
+def bootstrap_callback():
+    pass
