+++ config/services_demo_onion.py
@@ -1,0 +1,9 @@
+from os.path import join, split
+import sys
+
+service_to_command = {
+    "demo.onion": [
+        sys.executable,
+        join(split(__file__)[0], 'ns_always_txtorcon.py'),
+    ],
+}
